fix milvus search data return NoneType, fix unstructured loader when the directory with result does not exist, update README Configuration with VectorDB, FileLoader, WebCrawler, adding examples for loading local file and website (#24)
* refined unstructured loader to implement single file loading

* refined unstructured loader to implement single file loading

* fix milvus search data return NoneType, fix unstructured loader when the directory with result does not exist, update README Configuration with VectorDB, FileLoader, WebCrawler, adding examples for loading local file and website

* fix README Configuration typo

+++ deepsearcher/loader/file_loader/unstructured_loader.py
@@ -8,7 +8,8 @@ from deepsearcher.tools import log
 class UnstructuredLoader(BaseLoader):
     def __init__(self):
         self.directory_with_results = "./pdf_processed_outputs"
-        shutil.rmtree(self.directory_with_results)
+        if os.path.exists(self.directory_with_results):
+            shutil.rmtree(self.directory_with_results)
         os.makedirs(self.directory_with_results)
 
     def load_pipeline(self, input_path: str) -> List[Document]:
